requirement.txt
certifi==2021.10.8
chardet==4.0.0
Django==1.11.29
django-widget-tweaks==1.4.11
djangorestframework==3.9.4
djangorestframework-simplejwt==4.1.4
idna==2.10
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2025.2
requests==2.27.1
urllib3==1.26.20
